+++ release-notes.md
@@ -1,4 +1,15 @@
 # Propus 16.7.0  #
+## Add more user details name to Q&A response profile  ##
+- Issue Type: Feature
+- Issue ID: WEBC-1925
+
+### Configuration ###
+    First replace all tokens from the XML files below and remove ".template" from the file name:
+    /opt/kaltura/app/deployment/updates/scripts/xml/2020_07_16_updateQandAResponseProfile_addMoreUserData.template.xml
+		
+#### Deployment Scripts ####	
+    php /opt/kaltura/app/deployment/updates/scripts/2020_07_16_updateQandAResponseProfile_addMoreUserData.php
+
 ## Adding permission to serve and getServeUrl in entryVendorTask ##
 Issue Type: Task
 Issue ID : REACH2-879